Debugged an issue where the program was effectively blowing up as it wasn't waiting long enough after clicking load more to start populating the lists used to form the csv file of collected data. SHOULD be fixed... but, I'm not convinced it's a surefire fix.
At this point the filters seem to work, for singular and multiple terms, as well as for pages that may not have enough results for a load more button to be present. Or, searches that yield no results at all.

Pushing of the data to a csv file seems to be solved now, may need to clean pricing data, although maybe just find a way to get discounted prices down the line.