Fixed a bug with input display for searching (Specifically the output for term being searched for wouldn't immediately be a clickable link that was correct as search queries with spaces in them would yield an incomplete link)
Ex: User searches for Khaki pants
Display says: Currently printing info for https://www.asos.com/us/search/?q=khaki pants hence not all one link, update it should now be
                       Currently printing info for https://www.asos.com/us/search/?q=shirts%20pants which SHOULD be the correct search link. Small gripe but, EH.

Also fixed a bug in reference to stale element exception, that is, for example, pages that don't have enough products present to generate a load more button, had to add an exception to handle when the program goes to click the button but it's not there in the first place. Should work now.